add provenance scss (and handle overflow), refactor layout into utils
