+++ src/lib/containers/provenance/ProvenanceGraph.tsx
@@ -4,17 +4,16 @@ import ReactFlow, {
   Controls,
   Node,
   Edge,
-  Position,
   ConnectionLineType,
   useNodesState,
   useEdgesState,
 } from 'react-flow-renderer'
 import {
+  getLayoutedElements,
   getProvenanceEdge,
   getProvenanceNode,
   NodeType,
 } from './ProvenanceUtils'
-import dagre from 'dagre'
 import useDeepCompareEffect from 'use-deep-compare-effect'
 import {
   Used,
@@ -104,48 +103,6 @@ export const ProvenanceGraph = (props: ProvenanceProps) => {
     }
   }, [nodes, edges, rootActivity, rootNodeProps, setNodes, setEdges])
 
-  // layout
-  const dagreGraph = new dagre.graphlib.Graph()
-  const nodeWidth = 172
-  const nodeHeight = 36
-  dagreGraph.setDefaultEdgeLabel(() => ({}))
-
-  const getLayoutedElements = (
-    nodes: Node[],
-    edges: Edge[],
-    direction: string,
-  ) => {
-    const isHorizontal = direction === 'LR'
-    dagreGraph.setGraph({ rankdir: direction })
-
-    nodes.forEach(node => {
-      dagreGraph.setNode(node.id, { width: nodeWidth, height: nodeHeight })
-    })
-
-    edges.forEach(edge => {
-      dagreGraph.setEdge(edge.source, edge.target)
-    })
-
-    dagre.layout(dagreGraph)
-
-    nodes.forEach(node => {
-      const nodeWithPosition = dagreGraph.node(node.id)
-      node.targetPosition = isHorizontal ? Position.Left : Position.Top
-      node.sourcePosition = isHorizontal ? Position.Right : Position.Bottom
-
-      // We are shifting the dagre node position (anchor=center center) to the top left
-      // so it matches the React Flow node anchor point (top left).
-      node.position = {
-        x: nodeWithPosition.x - nodeWidth / 2,
-        y: nodeWithPosition.y - nodeHeight / 2,
-      }
-
-      return node
-    })
-
-    return { nodes, edges }
-  }
-
   useDeepCompareEffect(() => {
     const { nodes: layoutedNodes, edges: layoutedEdges } = getLayoutedElements(
       nodes,
@@ -155,6 +112,7 @@ export const ProvenanceGraph = (props: ProvenanceProps) => {
     setLayoutedNodes(layoutedNodes)
     setLayoutedEdges(layoutedEdges)
   }, [nodes, edges])
+
   return (
     <div
       className="bootstrap-4-backport ProvenanceWidget"
