+++ src/lib/style/components/_all.scss
@@ -20,4 +20,5 @@
   './blocking-loader', './transition', './popover', './featured-tool-card',
   './row_renderers/observation-card', './subsection-row-renderer',
   './dataset-editor', './fluid-modal', './markdown-collapse',
-  './discussion-search', './submission-page', './governance-dashboard';
+  './discussion-search', './submission-page', './governance-dashboard',
+  './provenance';
