+++ src/lib/style/components/_provenance.scss
@@ -1,0 +1,4 @@
+.react-flow__node {
+  text-overflow: ellipsis;
+  overflow: clip;
+}
